// src/types/clinic.d.ts
	interface IVaccinationRecord {
		_id: string;
		type: 'VACCINATION';
		uploadedBy: 'CLINIC';
		url: string | null;
		comment: string | null;
		vaccinationDate: string;
		nextVaccinationDate: string | null;
		vaccinatedOnDate: string;
		followUpDate: string | null;
		followUpType: string | null;
		vaccineName: string;
		notificationCount: number;
		active: boolean;
		parent: ICommonTypes.IParent;
		clinic: ICommonTypes.IClinic;
		pet: ICommonTypes.IPet;
		createdAtIST: string;
		updatedAtIST: string;
		recordId: string;
		createdAtUTC: string;
		updatedAtUTC: string;
	}
// ...
	interface IParent {
		name: string;
		mobile: string;
		parentId: string;
	}